pyproject.toml
[tool.poetry]
name = "rest-filefield"
version = "0.1.0"
description = ""
authors = ["Nanashi <alksndr.zln@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.18"
environs = "^9.5.0"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
django-rest-framework = "^0.1.0"
celery = "^5.3.1"
redis = "^4.5.5"
drf-spectacular = "^0.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
